refactor: 해당 View 와 관련된 파일들을 같은 패키지에 위치
