+++ app/src/main/java/woowacourse/movie/fragment/movielist/MovieListFragment.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.activity.MovieReservationActivity
-import woowacourse.movie.view.adapter.MovieAdapter
+import woowacourse.movie.fragment.movielist.adapter.MovieAdapter
 import woowacourse.movie.view.data.MovieListViewData
 import woowacourse.movie.view.data.MovieListViewType
 import woowacourse.movie.view.data.MovieViewData
