+++ app/src/main/java/woowacourse/movie/fragment/movielist/adapter/AdvertisementViewHolder.kt
@@ -1,9 +1,9 @@
-package woowacourse.movie.view.viewholder
+package woowacourse.movie.fragment.movielist.adapter
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class AdvertisementViewHolder(private val view: View, onClickItem: (Int) -> Unit) :
+class AdvertisementViewHolder(view: View, onClickItem: (Int) -> Unit) :
     RecyclerView.ViewHolder(view) {
     init {
         view.setOnClickListener {
